requirements.txt
cffi==1.14.5
click==8.0.0
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.17.1
python-slugify==5.0.2
text-unidecode==1.3
Werkzeug==2.0.1
